requirements.txt
pytest == 7.4.3
requests == 2.32.4
allure-pytest == 2.13.2
pytest-xdist == 3.3.1
pytest-rerunfailures == 12.0
PyMySQL == 1.1.0
redis == 4.5.5
pika == 1.3.2
pycryptodome == 3.20.0
cryptography == 44.0.1
PyYAML == 6.0
python-dotenv == 1.0.0
jsonschema == 4.19.0
python-json-logger == 2.0.7